requirements.txt
pillow==6.2.1
pytorch==1.2.0
torchvision==0.4.0
cudatoolkit==10.0.130
opencv==3.4.7
setuptools==39.1.0
scikit-learn==0.21.3
